refactor the code into better packages, improve the delete token dialog
