--- src/main/java/uk/co/bitethebullet/android/token/IToken.java
@@ -1,39 +1,0 @@
-/*
- * Copyright Mark McAvoy - www.bitethebullet.co.uk 2009
- * 
- * This file is part of Android Token.
- *
- * Android Token is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Android Token is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Android Token.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-package uk.co.bitethebullet.android.token;
-
-public interface IToken {
-	
-	public String getName();
-	
-	public String getSerialNumber();
-	
-	public int getTokenType();
-	
-	public String generateOtp();
-	
-	public long getId();
-	
-	public void setId(long id);
-	
-	public int getTimeStep();
-
-	public String getOrganisation();
-}
