--- src/main/java/uk/co/bitethebullet/android/token/ITokenMeta.java
@@ -1,18 +1,0 @@
-package uk.co.bitethebullet.android.token;
-
-public interface ITokenMeta {
-
-	public String getName();
-
-	public String getOrganisation();
-	
-	public int getTokenType();
-	
-	public String getSecretBase32();
-	
-	public int getDigits();
-	
-	public int getTimeStep();
-	
-	public int getCounter();
-}
