--- src/main/java/uk/co/bitethebullet/android/token/OtpAuthUriException.java
@@ -1,24 +1,0 @@
-package uk.co.bitethebullet.android.token;
-
-public class OtpAuthUriException extends Exception {
-	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	String err;
-	
-	public OtpAuthUriException()
-	  {
-	    super();             // call superclass constructor
-	    err = "unknown";
-	  }
-	  
-	//-----------------------------------------------
-	// Constructor receives some kind of message that is saved in an instance variable.
-	  public OtpAuthUriException(String err)
-	  {
-	    super(err);     // call super class constructor
-	    this.err = err;  // save message
-	  }
-}
